requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
greenlet
websockets
python-dotenv
jinja2
